#180 - 리뷰 인기 점수 관리 메소드 분리
