+++ src/main/java/com/ncookie/imad/domain/ranking/service/TodayPopularScheduler.java
@@ -17,14 +17,14 @@ public class TodayPopularScheduler {
     @Description("매일 자정마다 인기 리뷰 데이터 리셋")
     @Scheduled(cron = "0 0 0 * * ?")    // 자정마다 실행
     public void clearPopularReviewDaily() {
-        todayPopularScoreService.clearDaily();
+        todayPopularScoreService.clearPopularReviewScore();
         log.info("오늘의 리뷰 데이터 리셋 완료");
     }
 
     @Description("매일 자정마다 인기 게시글 데이터 리셋")
     @Scheduled(cron = "0 0 0 * * ?")    // 자정마다 실행
     public void clearPopularPostingDaily() {
-        todayPopularScoreService.clearDaily();
+        todayPopularScoreService.clearPopularPostingScore();
         log.info("오늘의 게시글 데이터 리셋 완료");
     }
 }
